b - gcc complains that ssize_t is not declared/defined and, til, ssize_t is not part of the C++ standard. It is part of POSIX, however, and used in read(2) and compilers have made a habit of defining it, as clang apparently still does.

+++ include/readseek.h
@@ -1,13 +1,14 @@
 #pragma once
 
 #include <cstddef>
+#include <cstdint>
 #include <vector>
 
 class ReadSeek {
 public:
   virtual ~ReadSeek() {}
 
-  virtual ssize_t read(size_t len, std::vector<uint8_t>& buf) = 0;
+  virtual int64_t read(size_t len, std::vector<uint8_t>& buf) = 0;
 
   virtual size_t tellg() const = 0;
   virtual size_t seek(size_t pos) = 0;
